+++ rosetta-source/src/main/resources/enrich/config/test-pack-enrich-vehicle-ownership-to-reportable-vehicle-passenger-vehicles.json
@@ -5,7 +5,7 @@
   "samples" : [ {
     "id" : "demo-1",
     "name" : "demo 1",
-    "inputPath" : "enrich/input/passenger-vehicles/demo-1.json",
+    "inputPath" : "ingest/output/external-vehicle-and-licence-data-to-vehicle-ownership/passenger-vehicles/demo-1.json",
     "outputPath" : "enrich/output/vehicle-ownership-to-reportable-vehicle/passenger-vehicles/demo-1.json",
     "assertions" : {
       "modelValidationFailures" : 3,
@@ -14,7 +14,7 @@
   }, {
     "id" : "demo-2",
     "name" : "demo 2",
-    "inputPath" : "enrich/input/passenger-vehicles/demo-2.json",
+    "inputPath" : "ingest/output/external-vehicle-and-licence-data-to-vehicle-ownership/passenger-vehicles/demo-2.json",
     "outputPath" : "enrich/output/vehicle-ownership-to-reportable-vehicle/passenger-vehicles/demo-2.json",
     "assertions" : {
       "modelValidationFailures" : 3,
