--- rosetta-source/src/main/resources/enrich/input/passenger-vehicles/demo-1.json
@@ -1,59 +1,0 @@
-{
-  "drivingLicence": [
-    {
-      "person": {
-        "firstName": "Joe",
-        "surname": "Bloggs"
-      },
-      "countryOfIssuance": {
-        "value": "GB"
-      },
-      "licenceNumber": "AB12CDE",
-      "dateOfIssuance": "2019-07-01",
-      "dateOfRenewal": "2029-07-01",
-      "vehicleEntitlement": [ "M_1_PASSENGERS", "M_2_PASSENGERS", "M_3_PASSENGERS" ],
-      "penaltyPoints": 3
-    }
-  ],
-  "vehicle" : {
-    "CombustibleVehicle": {
-      "Car": {
-        "specification": {
-          "make": "Volkswagen",
-          "model": "Polo",
-          "carType": "HATCHBACK",
-          "zeroTo60": 4.5,
-          "yearOfManufacture": 2020,
-          "chassisNumber": 0,
-          "dateOfFirstRegistration": {
-            "year": 2020,
-            "month": 1,
-            "day": 1
-          }
-        },
-        "registrationID": "X2",
-        "vehicleTaxBand": "B_BAND",
-        "vehicleClassification": "M_1_PASSENGERS",
-        "fuelCapacity": 200,
-        "engine": {
-          "engineType": "PETROL",
-          "power": 1,
-          "mpg": 1,
-          "emissionMetrics": {
-            "carbonMonoxide" : 2.70,
-            "totalHydrocarbon" : 0,
-            "volatileOrganicCompound" : 0,
-            "nitrogenOxide" : 0,
-            "combineHCNOX" : 0,
-            "particulateMatter" : 0,
-            "particleNumber" : 0
-          }
-        },
-        "carType": "HATCHBACK"
-      }
-    }
-  },
-  "dateOfPurchase": "2020-01-02",
-  "isFirstHand": true,
-  "vehicleOwnershipRegistrationID": 1
-}
