--- rosetta-source/src/main/resources/enrich/input/passenger-vehicles/demo-2.json
@@ -1,62 +1,0 @@
-{
-  "drivingLicence": [
-    {
-      "person": {
-        "firstName": "Jane",
-        "surname": "Doe"
-      },
-      "countryOfIssuance": {
-        "value": "GB"
-      },
-      "licenceNumber": "FG34HJK",
-      "dateOfIssuance": "2018-08-01",
-      "dateOfRenewal": "2028-08-01",
-      "vehicleEntitlement": [ "M_1_PASSENGERS", "M_2_PASSENGERS", "M_3_PASSENGERS" ],
-      "penaltyPoints": 0
-    }
-  ],
-  "vehicle" : {
-    "CombustibleVehicle": {
-      "Car": {
-        "specification": {
-          "make": "Volkswagen",
-          "model": "Golf",
-          "zeroTo60": 4.5,
-          "yearOfManufacture": 2020,
-          "chassisNumber": 0,
-          "dateOfFirstRegistration": {
-            "year": 2020,
-            "month": 6,
-            "day": 1
-          }
-        },
-        "registrationID": "X1",
-        "vehicleTaxBand": "A_BAND",
-        "vehicleClassification": "M_2_PASSENGERS",
-        "fuelCapacity": 150,
-        "engine": {
-          "engineType": "PETROL",
-          "power": 1,
-          "mpg": 1,
-          "emissionMetrics": {
-            "carbonMonoxide" : 2.66,
-            "totalHydrocarbon" : 0,
-            "volatileOrganicCompound" : 0,
-            "nitrogenOxide" : 0,
-            "combineHCNOX" : 0,
-            "particulateMatter" : 0,
-            "particleNumber" : 0
-          }
-        },
-        "carType": "HATCHBACK"
-      }
-    }
-  },
-  "dateOfPurchase": {
-    "year": 2020,
-    "month": 6,
-    "day": 2
-  },
-  "isFirstHand": true,
-  "vehicleOwnershipRegistrationID": 1
-}
