+++ rosetta-source/src/main/resources/ingest/config/test-pack-ingest-external-vehicle-and-licence-data-to-vehicle-ownership-passenger-vehicles.json
@@ -1,0 +1,24 @@
+{
+  "id" : "test-pack-ingest-external-vehicle-and-licence-data-to-vehicle-ownership-passenger-vehicles",
+  "pipelineId" : "pipeline-ingest-external-vehicle-and-licence-data-to-vehicle-ownership",
+  "name" : "Passenger Vehicles",
+  "samples" : [ {
+    "id" : "demo-1",
+    "name" : "demo 1",
+    "inputPath" : "ingest/input/passenger-vehicles/demo-1.json",
+    "outputPath" : "ingest/output/external-vehicle-and-licence-data-to-vehicle-ownership/passenger-vehicles/demo-1.json",
+    "assertions" : {
+      "modelValidationFailures" : 3,
+      "runtimeError" : false
+    }
+  }, {
+    "id" : "demo-2",
+    "name" : "demo 2",
+    "inputPath" : "ingest/input/passenger-vehicles/demo-2.json",
+    "outputPath" : "ingest/output/external-vehicle-and-licence-data-to-vehicle-ownership/passenger-vehicles/demo-2.json",
+    "assertions" : {
+      "modelValidationFailures" : 3,
+      "runtimeError" : false
+    }
+  } ]
+}
