+++ rosetta-source/src/main/resources/ingest/input/passenger-vehicles/demo-1.json
@@ -1,0 +1,43 @@
+{
+  "vehicleData" : {
+    "vehicleType": "CAR",
+    "dateOfPurchase": "2020-01-02",
+    "isFirstHand": true,
+    "vehicleOwnershipRegistrationID": 1,
+    "make": "Volkswagen",
+    "model": "Polo",
+    "zeroTo60": 4.5,
+    "yearOfManufacture": 2020,
+    "chassisNumber": 0,
+    "dateOfFirstRegistration": "2020-01-01",
+    "registrationID": "X2",
+    "vehicleTaxBand": "BBand",
+    "vehicleClassification": "M1_Passengers",
+    "engineType": "Petrol",
+    "power": 1,
+    "mpg": 1,
+    "carbonMonoxide" : 2.70,
+    "totalHydrocarbon" : 0,
+    "volatileOrganicCompound" : 0,
+    "nitrogenOxide" : 0,
+    "combineHCNOX" : 0,
+    "particulateMatter" : 0,
+    "particleNumber" : 0,
+    "carType": "Hatchback",
+    "fuelCapacity": 200
+  },
+  "drivingLicenceData": [
+    {
+      "firstName": "Joe",
+      "surname": "Bloggs",
+      "countryOfIssuance": {
+        "value": "GB"
+      },
+      "licenceNumber": "AB12CDE",
+      "dateOfIssuance": "2019-07-01",
+      "dateOfRenewal": "2029-07-01",
+      "vehicleEntitlement": [ "M1_Passengers", "M2_Passengers", "M3_Passengers" ],
+      "penaltyPoints": 3
+    }
+  ]
+}
