+++ rosetta-source/src/main/resources/ingest/output/external-vehicle-and-licence-data-to-vehicle-ownership/passenger-vehicles/demo-1.json
@@ -1,0 +1,52 @@
+{
+  "dateOfPurchase" : "2020-01-02",
+  "drivingLicence" : [ {
+    "countryOfIssuance" : {
+      "value" : "GB"
+    },
+    "dateOfIssuance" : "2019-07-01",
+    "dateOfRenewal" : "2029-07-01",
+    "licenceNumber" : "AB12CDE",
+    "penaltyPoints" : 3,
+    "person" : {
+      "firstName" : "Joe",
+      "surname" : "Bloggs"
+    },
+    "vehicleEntitlement" : [ "M1_Passengers", "M2_Passengers", "M3_Passengers" ]
+  } ],
+  "isFirstHand" : true,
+  "vehicle" : {
+    "CombustibleVehicle" : {
+      "Car" : {
+        "carType" : "Hatchback",
+        "engine" : {
+          "emissionMetrics" : {
+            "carbonMonoxide" : 2.70,
+            "combineHCNOX" : 0,
+            "nitrogenOxide" : 0,
+            "particleNumber" : 0,
+            "particulateMatter" : 0,
+            "totalHydrocarbon" : 0,
+            "volatileOrganicCompound" : 0
+          },
+          "engineType" : "Petrol",
+          "mpg" : 1,
+          "power" : 1
+        },
+        "fuelCapacity" : 200,
+        "registrationID" : "X2",
+        "specification" : {
+          "chassisNumber" : 0,
+          "dateOfFirstRegistration" : "2020-01-01",
+          "make" : "Volkswagen",
+          "model" : "Polo",
+          "yearOfManufacture" : 2020,
+          "zeroTo60" : 4.5
+        },
+        "vehicleClassification" : "M1_Passengers",
+        "vehicleTaxBand" : "BBand"
+      }
+    }
+  },
+  "vehicleOwnershipRegistrationID" : 1
+}
