+++ tests/src/test/java/demo/emissions/EmissionsIngestTest.java
@@ -1,10 +1,10 @@
-package demo.emissions.projection;
+package demo.emissions;
 
 import com.regnosys.rosetta.common.transform.TestPackModel;
 import com.regnosys.rosetta.common.transform.TransformType;
 import com.regnosys.testing.transform.TransformTestExtension;
-import demo.emissions.DemoTestRuntimeModule;
-import demo.emissions.projection.functions.Project_EuropeanParliamentReportToIsoEmissionsReport;
+import demo.emissions.enrichment.functions.Enrich_VehicleOwnershipToReportableVehicle;
+import demo.emissions.ingestion.functions.Ingest_ExternalVehicleAndLicenceDataToVehicleOwnership;
 import org.eclipse.xtext.testing.InjectWith;
 import org.eclipse.xtext.testing.extensions.InjectionExtension;
 import org.junit.jupiter.api.TestInstance;
@@ -21,22 +21,22 @@ import java.util.stream.Stream;
 @TestInstance(TestInstance.Lifecycle.PER_CLASS)
 @ExtendWith(InjectionExtension.class)
 @InjectWith(DemoTestRuntimeModule.InjectorProvider.class)
-class EmissionsProjectionTest {
+class EmissionsIngestTest {
 
-    public static final Path PROJECTION_CONFIG_PATH = Paths.get(TransformType.PROJECTION.getResourcePath()).resolve("config");
+    public static final Path INGEST_CONFIG_PATH = Paths.get(TransformType.TRANSLATE.getResourcePath()).resolve("config");
 
     @RegisterExtension
-    static TransformTestExtension<Project_EuropeanParliamentReportToIsoEmissionsReport> testExtension =
+    static TransformTestExtension<Ingest_ExternalVehicleAndLicenceDataToVehicleOwnership> testExtension =
             new TransformTestExtension<>(new DemoTestRuntimeModule(),
-                    PROJECTION_CONFIG_PATH,
-                    Project_EuropeanParliamentReportToIsoEmissionsReport.class);
+                    INGEST_CONFIG_PATH,
+                    Ingest_ExternalVehicleAndLicenceDataToVehicleOwnership.class);
 
     @ParameterizedTest(name = "{0}")
     @MethodSource("inputFiles")
     void runTest(String testName,
                  String testPackId,
                  TestPackModel.SampleModel sampleModel,
-                 Project_EuropeanParliamentReportToIsoEmissionsReport func) {
+                 Ingest_ExternalVehicleAndLicenceDataToVehicleOwnership func) {
         testExtension.runTransformAndAssert(testPackId, sampleModel, func::evaluate);
     }
