+++ tests/src/test/java/demo/emissions/EmissionsProjectionTest.java
@@ -1,10 +1,9 @@
-package demo.emissions.enrichment;
+package demo.emissions;
 
 import com.regnosys.rosetta.common.transform.TestPackModel;
 import com.regnosys.rosetta.common.transform.TransformType;
 import com.regnosys.testing.transform.TransformTestExtension;
-import demo.emissions.DemoTestRuntimeModule;
-import demo.emissions.enrichment.functions.Enrich_VehicleOwnershipToReportableVehicle;
+import demo.emissions.projection.functions.Project_EuropeanParliamentReportToIsoEmissionsReport;
 import org.eclipse.xtext.testing.InjectWith;
 import org.eclipse.xtext.testing.extensions.InjectionExtension;
 import org.junit.jupiter.api.TestInstance;
@@ -21,22 +20,22 @@ import java.util.stream.Stream;
 @TestInstance(TestInstance.Lifecycle.PER_CLASS)
 @ExtendWith(InjectionExtension.class)
 @InjectWith(DemoTestRuntimeModule.InjectorProvider.class)
-class EmissionsEnrichTest {
+class EmissionsProjectionTest {
 
-    public static final Path ENRICH_CONFIG_PATH = Paths.get(TransformType.ENRICH.getResourcePath()).resolve("config");
+    public static final Path PROJECTION_CONFIG_PATH = Paths.get(TransformType.PROJECTION.getResourcePath()).resolve("config");
 
     @RegisterExtension
-    static TransformTestExtension<Enrich_VehicleOwnershipToReportableVehicle> testExtension =
+    static TransformTestExtension<Project_EuropeanParliamentReportToIsoEmissionsReport> testExtension =
             new TransformTestExtension<>(new DemoTestRuntimeModule(),
-                    ENRICH_CONFIG_PATH,
-                    Enrich_VehicleOwnershipToReportableVehicle.class);
+                    PROJECTION_CONFIG_PATH,
+                    Project_EuropeanParliamentReportToIsoEmissionsReport.class);
 
     @ParameterizedTest(name = "{0}")
     @MethodSource("inputFiles")
     void runTest(String testName,
                  String testPackId,
                  TestPackModel.SampleModel sampleModel,
-                 Enrich_VehicleOwnershipToReportableVehicle func) {
+                 Project_EuropeanParliamentReportToIsoEmissionsReport func) {
         testExtension.runTransformAndAssert(testPackId, sampleModel, func::evaluate);
     }
